% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_blank.R
\name{axis_blank}
\alias{axis_blank}
\title{Add an axis to a plot without labels as the default}
\usage{
axis_blank(
  side,
  at = NULL,
  labels = FALSE,
  tick = TRUE,
  line = NA,
  pos = NA,
  outer = FALSE,
  font = NA,
  lty = "solid",
  lwd = 1,
  lwd.ticks = lwd,
  col = NULL,
  col.ticks = NULL,
  hadj = NA,
  padj = NA,
  gap.axis = NA,
  ...,
  tck = -0.02,
  minor = TRUE
)
}
\arguments{
\item{side}{an integer specifying which side of the plot the axis is to be drawn on. The axis is placed as follows: 1=below, 2=left, 3=above and 4=right.}

\item{at}{The points at which tick-marks are to be drawn. Non-finite (infinite, NaN or NA) values are omitted. By default (when NULL) tickmark locations are computed using \code{\link[graphics]{axTicks}}.}

\item{labels}{Set to \code{FALSE}. See \code{\link[graphics]{axis}}.}

\item{tick}{See \code{\link[graphics]{axis}}.}

\item{line}{See \code{\link[graphics]{axis}}.}

\item{pos}{See \code{\link[graphics]{axis}}.}

\item{outer}{See \code{\link[graphics]{axis}}.}

\item{font}{See \code{\link[graphics]{axis}}.}

\item{lty}{See \code{\link[graphics]{axis}}.}

\item{lwd}{See \code{\link[graphics]{axis}}.}

\item{lwd.ticks}{See \code{\link[graphics]{axis}}.}

\item{col}{See \code{\link[graphics]{axis}}.}

\item{col.ticks}{See \code{\link[graphics]{axis}}.}

\item{hadj}{See \code{\link[graphics]{axis}}.}

\item{padj}{See \code{\link[graphics]{axis}}.}

\item{gap.axis}{See \code{\link[graphics]{axis}}.}

\item{...}{Other graphical parameters that may apply to \code{\link[graphics]{axis}}.}

\item{tck}{The length of tick marks as a fraction of the smaller of the width or height of the plotting region. If tck >= 0.5 it is interpreted as a fraction of the relevant side, so if tck = 1 grid lines are drawn. The default setting is (tck = -0.025).}

\item{minor}{Whether or not to add smaller tick marks spaced equally between the larger tickmarks specified by the at argument. Tick length is set to \code{tck * 0.5}.}
}
\description{
A wrapper function for \code{\link[base]{axis}}
  with three changes.
   
  1. \code{labels = FALSE} is now the default.
  2. \code{tck} has been added as an argument, which is used to specify the size of the tick mark.
  3. \code{minor} has been added as an argument. Used to add minor (i.e., smaller) tickmarks equally between the first axis.
}
\examples{
\dontrun{
   blank(
     xlim = c(0,50),
     ylim = c(0,100),
     bty ="l"
   )
   
   axis_blank(1, at = seq(0,50,10))
   axis_blank(2, at = seq(0,100,20))
   
}

}
