% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_text.R
\name{axis_text}
\alias{axis_text}
\title{Add labels to the axis}
\usage{
axis_text(
  text = NA,
  side = 1,
  line = 0.75,
  outer = FALSE,
  at = NA,
  adj = NA,
  padj = NA,
  cex = 1.2,
  col = NA,
  font = NA,
  ...
)
}
\arguments{
\item{text}{a character or expression vector specifying the text to be written.
If NA, then \code{\link[graphics]{axTicks}} is used to find values.}

\item{side}{an integer specifying which side of the plot the axis is to be drawn on. The axis is placed as follows: 1=below, 2=left, 3=above and 4=right.}

\item{line}{at which MARgin line, starting at 0 counting outwards. Defaults to 0.75.}

\item{outer}{use outer margins if available}

\item{at}{give location of each string in user coordinates. If \code{text} can be coerced to numeric values, it will try those. Otherwise, if the component of at corresponding to a particular text item is not a finite value (the default), the location will be determined by \code{adj.}}

\item{adj}{set to NA. See \code{\link[graphics]{mtext}}.}

\item{padj}{set to NA. See \code{\link[graphics]{mtext}}.}

\item{cex}{character expansion factor. Defaults to 1.2. See \code{\link[graphics]{mtext}} for more details.}

\item{col}{color to use.  Can be a vector. \code{NA} values (the default) mean use \code{par("col")}.}

\item{font}{font for text. Can be a vector. \code{NA} values (the default) mean use \code{par("font")}.}

\item{...}{Other graphical parameters that may apply to \code{\link[graphics]{mtext}}.}
}
\description{
A wrapper function for \code{\link[base]{mtext}}. 
The changes made to this function are simple quality of life improvments. For
example, if \code{text = NA}, then the function uses \code{\link[graphics]{axTicks}}
to locate what values to input.
}
\examples{
\dontrun{
   blank(
     xlim = c(0,50),
     ylim = c(0,100),
     bty ="l"
   )
   
   axis_blank(side = 1, at = seq(0,50,10))
   axis_blank(side = 2, at = seq(0,100,20))
   axis_text(side = 1)
   axis_text(side = 2)
   
}

}
