% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blank.R
\name{blank}
\alias{blank}
\title{Create a plot with a blank space as the default}
\usage{
blank(
  x = 1,
  y = 1,
  type = "n",
  xlab = "",
  ylab = "",
  xaxt = "n",
  yaxt = "n",
  bty = "n",
  ...
)
}
\arguments{
\item{x}{Set to \code{1}. See \code{\link[base]{plot}}.}

\item{y}{Set to \code{1}. See \code{\link[base]{plot}}.}

\item{type}{Set to \code{"n"}. See \code{\link[base]{plot}}.}

\item{xlab}{Set to \code{""}. See \code{\link[base]{plot}}.}

\item{ylab}{Set to \code{""}. See \code{\link[base]{plot}}.}

\item{xaxt}{Set to \code{"n"}. See \code{\link[base]{plot}}.}

\item{yaxt}{Set to \code{"n"}. See \code{\link[base]{plot}}.}

\item{bty}{Set to \code{"n"}. See \code{\link[base]{plot}}.}

\item{...}{Additional arguments sent to \code{\link[base]{plot}}.}
}
\description{
A simple wrapper on \code{\link[base]{plot}} with all
  default values set to create a blank plot space. \cr \cr
  Specifically, \code{x = 1}, \code{y = 1}, \code{type = "n"}, 
  \code{xlab = ""}, \code{ylab = ""}, \code{xaxt = "n"}, \code{yaxt = "n"}, 
  and \code{bty = "n"}. Any of the input values can be overwritten within
  this function's call.
}
\examples{
\dontrun{
  blank()   
}

}
