% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violin_functions.R
\name{dist_values}
\alias{dist_values}
\title{determines the values of the variable to evaluate (x_values) and the 
resulting evaluation values (y_values) of the distribution. these are 
the raw values}
\usage{
dist_values(x, values = NULL, nvalues = NULL, type = "l")
}
\value{
data.frame of x and y
}
\description{
determines the values of the variable to evaluate (x_values) and the 
resulting evaluation values (y_values) of the distribution. these are 
the raw values
}
