% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssw_style.R
\name{ssw_style}
\alias{ssw_style}
\title{Add ssw theme to ggplot chart}
\usage{
ssw_style()
}
\description{
This function allows you to add the ssw theme to your ggplotgraphics.
}
\examples{
line <- ggplot(line_df, aes(x = year, y = lifeExp)) +
geom_line(colour = "#007f7f", size = 1) +
geom_hline(yintercept = 0, size = 1, colour="#333333") +
ssw_style()
}
\keyword{ssw_style}
