% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_text.R
\name{axis_text}
\alias{axis_text}
\title{Add labels to the axis labels}
\usage{
axis_text(
  text = NULL,
  labels = FALSE,
  side = 3,
  line = 0,
  outer = FALSE,
  at = NULL,
  adj = NA,
  padj = NA,
  cex = NA,
  col = NA,
  font = NA,
  ...
)
}
\arguments{
\item{text}{a character or expression vector specifying the text to be written. See \code{\link[graphics]{mtext}}.}

\item{labels}{Set to \code{FALSE}. See \code{\link[graphics]{axis}}.}

\item{side}{an integer specifying which side of the plot the axis is to be drawn on. The axis is placed as follows: 1=below, 2=left, 3=above and 4=right.}

\item{line}{on which MARgin line, starting at 0 counting outwards.}

\item{outer}{use outer margins if available.}

\item{at}{The location of each string in user coordinates (i.e., the text). See \code{\link[graphics]{mtext}}.}

\item{adj}{adjustment for each string in reading direction. For strings parallel to the axes, adj = 0 means left or bottom alignment, and adj = 1 means right or top alignment. See \code{\link[graphics]{mtext}}.}

\item{padj}{adjustment for each string perpendicular to the reading direction (which is controlled by \code{adj}). See \code{\link[graphics]{mtext}}.}

\item{cex}{character expansion factor. Can be a vector.}

\item{col}{color to use. Can be a vector. \code{NA} values (the default) means use \code{par("col")}.}

\item{font}{font for text. Can be a vector. \code{NA} values (the default) means use \code{par("font")}.}

\item{...}{Other graphical parameters that may apply to \code{\link[graphics]{mtext}}.}
}
\description{
A wrapper function for \code{\link[graphics]{mtext}}
  with two changes.
   
  1. The default for \code{text} and \code{at} is now \code{NULL}. If both are \code{NULL}, \code{\link[graphics]{axTicks}} will be used to determine what values to include on the x (\code{side = 1}) or y (\code{side = 2})  axis. 
  2. If \code{length(text) == 1} and \code{at = NULL}, then the appropriate \code{side} will find the central point of the axis to put the text (e.g., for an axis title).
}
\examples{
\dontrun{
   blank(
     xlim = c(0,50),
     ylim = c(0,100),
     bty ="l"
   )
   
   axis_blank(1, at = seq(0,50,10))
   axis_text(side = 1)
   axis_blank(2, at = seq(0,100,20))
   axis_text(side = 2)
   
}

}
