% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violin_functions.R
\name{violin}
\alias{violin}
\title{Draw violins}
\usage{
violin(
  x,
  location = NULL,
  rotate = TRUE,
  type = NULL,
  wex = 1,
  values = NULL,
  nvalues = NULL,
  side = "both",
  ...
)
}
\arguments{
\item{x}{The vector of values or function to be summarized}

\item{location}{The graphical location with respect to the axes where
the violin is to be placed. Default assumption is for the x-axis, but
can be a named-by-axis vector.}

\item{type}{Character indicating the type of plotting. Defaults to NULL
which defines itself then based on the support for x to 
either "l" (line) for continuous and "r" (histogram-like rectangles) for 
non-continuous/integer-conformable}

\item{wex}{Numeric height scale that transforms the distribution
evalution to the plotting axis value. now wex for width expansion}

\item{nvalues}{Integer number of values to use for the drawing of the 
violin. If NULL, set by vtype.}
}
\description{
Draw violins on plots
}
